dcdw/requirements.txt
boto3
elasticsearch-dsl>=6.0.0,<7.0.0
flask
flask-cors
lxml
numpy
pandas
pika
git+https://github.com/aboSamoor/polyglot.git@master
psycopg2-binary
pycountry
pyyaml
redis
requests
sklearn
sqlalchemy
waitress